requirements.txt
pyyaml==6.0.1
osmnx
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
scikit-learn==1.3.1
geopy
flask